Refactor and split tests into own files
