+++ aquadoggo/src/proptests/mod.rs
@@ -1,7 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-pub mod document_strategies;
-pub mod filter_strategies;
-pub mod schema_strategies;
-mod tests;
+pub mod strategies;
+mod pagination;
+mod filtering;
 pub mod utils;
+
+pub use utils::queries;
